Generate web map:
- Add index.js file, which has a js function to initiate a web map.
- Currently the map only contains a basemap, which uses vector tiles
from MAPC and data from MassGIS.
- Add an empty <div> to index.html with an id of `map`. This id is
the same as the `map` element in index.js. Leaflet generates the map
within this empty div.
- Add temporary width and height to the map in the css file.
- Create scratch file, which contains the same js function that generates
the map, but with center and zoom written out explicitly. This could be
helpful to reference later on.

Fix bug that was preventing the vector tiles from rendering
properly. Resolve bug by adding stylesheet to link tags. This MUST
be taken out of the head. Next commit message should resolve this
through the following steps:
  1. download file
  2. add file to style folder
  3. import into scss file